create groupedInvestigations utils
